Add address masks to BSSID lists
In many applications it is useful not just to enumerate a group of well known
access points, but to use a address/mask notation to match an entire set of
addresses (ca:ff:ee:00:00:00/FF:FF:FF:00:00:00).

This change expands the data structures used by mac lists to include a mask
indicating the significant (non-masked) portions of an address and extends the
list parser to recognize mask suffixes.