Clone SSID twins to enable encrypted catchall APs
This change makes it possible to use encryption with catchall networks.
Upon associating a new client with a custom SSID, a new "twin" of the existing
ssid structure is created and keys are regenerated using the new SSID and the
existing origin configuration as a template. So multiple clones of the same
network can coexist using different SSID names - hence the term "twin".